media: Organize thumbnail implementation units.
