+++ src/service/media/tests.rs
@@ -454,9 +454,9 @@ mod animation {
 		codecs::gif::{GifDecoder, GifEncoder},
 	};
 
-	use super::{super::thumbnail::animate::encode_frames, scale};
+	use super::{super::thumbnail::encode_frames, scale};
 
-	/// Pixels the tests give the encoder to spend, past anything they need.
+	// Pixels the tests give the encoder to spend, past anything they need.
 	const BUDGET: u64 = 1_000_000;
 
 	/// A canvas past the budget is refused before any frame is decoded.
