+++ src/service/media/thumbnail/animate.rs
@@ -1,249 +1,177 @@
-//! Animated thumbnail encoding
-//!
-//! A picture that carries a frame sequence can answer a request for an animated
-//! thumbnail with the frames themselves rather than one of them. The three
-//! containers MSC2705 names each hold a sequence differently, so each has its
-//! own decoder here, and all of them produce a GIF, which is the one animated
-//! format this build can encode.
-//!
-//! The caps that bound the work live here rather than at either call site, so
-//! a source reaches no decoder before what it would cost is known.
+//! Defines animated thumbnail admission and selection policy.
 
-use std::{
-	cmp::min,
-	io::Cursor,
-	panic::{AssertUnwindSafe, catch_unwind},
+use tuwunel_core::implement;
+
+use super::{
+	super::Fetched,
+	sniff::{animated_type, animates},
 };
 
-use bytes::Bytes;
-use image::{
-	AnimationDecoder, DynamicImage, Frame, Frames, ImageDecoder, ImageFormat, Limits,
-	codecs::{
-		gif::{GifDecoder, GifEncoder, Repeat},
-		png::PngDecoder,
-		webp::WebPDecoder,
-	},
-};
-use ruma::Mxc;
-use tokio::sync::OwnedSemaphorePermit;
-use tuwunel_core::{Err, Error, Result, defer, err, implement};
-
-use super::{ANIMATED_NAME, BYTES_PER_PIXEL, Dim, GIF, Media, reader, thumbnail_generate};
-
-/// Quantization effort the GIF encoder spends per frame.
+/// Content types naming a container that can carry a frame sequence.
 ///
-/// The encoder's own default is 1, which buys quality at any cost; a thumbnail
-/// is small and any remote server can ask for one, so this trades a little of
-/// that for an encode that ends.
-const GIF_SPEED: i32 = 10;
+/// A picture read as animating is stored under the one its header names, so a
+/// later lookup holding the key rather than the picture still knows what the
+/// row carries.
+pub(super) const APNG: &str = "image/apng";
+pub(super) const GIF: &str = "image/gif";
+pub(super) const WEBP: &str = "image/webp";
 
-pub(super) struct Output {
-	pub(super) media: Result<Media>,
-	pub(super) admission: OwnedSemaphorePermit,
+/// Content types withheld from a request that asked for a still picture.
+///
+/// A still `image/webp` cannot be told from an animated one without decoding
+/// it, so the family is withheld whole. MSC2705 also names `image/png` for
+/// APNG, which cannot join the list because every generated thumbnail is one.
+pub(in super::super) const ANIMATED_TYPES: [&str; 3] = [APNG, GIF, WEBP];
+
+/// Whether a thumbnail request will accept an animated result.
+///
+/// MSC2705 gives the `animated` parameter three states and two behaviors: only
+/// `animated=true` may be answered with animation, while `animated=false` and
+/// an absent parameter alike forbid it.
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
+pub enum Animate {
+	/// The response must be a still picture.
+	#[default]
+	Never,
+
+	/// The response may animate.
+	Allowed,
+}
+/// Returns true when the request will accept animation.
+///
+/// Only `animated=true` reaches this state; `animated=false` and an absent
+/// parameter alike forbid animation.
+#[implement(Animate)]
+#[inline]
+#[must_use]
+pub fn allowed(self) -> bool { matches!(self, Self::Allowed) }
+
+/// Returns true when content of this type may answer the request at all.
+///
+/// This reads the declared type, which whoever uploaded the picture chose,
+/// so it is only for deciding between stored rows, where the pictures
+/// themselves are not in hand. Prefer [`Self::accepts_picture`] anywhere
+/// the bytes are.
+#[implement(Animate)]
+#[inline]
+#[must_use]
+pub fn accepts_type(self, content_type: Option<&str>) -> bool {
+	self.allowed() || !content_type.is_some_and(declares_animation)
 }
 
-/// Encode this picture's frames as an animated GIF thumbnail and store it.
+/// Returns true when this type is the variant the request asked for.
 ///
-/// Quantization runs per frame on a palette of its own, which is far more work
-/// than a still costs, so the encode is handed to a blocking worker rather than
-/// run on the async runtime. A source whose frames will not decode answers an
-/// error and the caller falls through to the still it would have produced.
-#[implement(super::super::Service)]
-#[tracing::instrument(
-	name = "animate",
-	level = "debug",
-	skip(self, source, admission)
-)]
-pub(super) async fn store_animated(
-	&self,
-	mxc: &Mxc<'_>,
-	dim: &Dim,
-	source: Bytes,
-	admission: OwnedSemaphorePermit,
-) -> Result<Output> {
-	let config = &self.services.config;
-	let max_frames = config.media_thumbnail_max_frames;
-	let budget = config.media_thumbnail_max_pixels;
-	let requested = Dim::new(dim.width, dim.height, Some(dim.method.clone()));
-
-	#[cfg(test)]
-	super::tests::before_animation_submit().await;
-	#[cfg(test)]
-	let control = super::tests::animation_control();
-
-	let encode = self
-		.services
-		.server
-		.runtime()
-		.spawn_blocking(move || {
-			#[cfg(test)]
-			if let Some(control) = control.as_ref() {
-				control.worker_started();
-			}
-
-			let content = catch_unwind(AssertUnwindSafe(|| {
-				encode_frames(&source, &requested, max_frames, budget)
-			}))
-			.map_err(Error::from_panic)
-			.unwrap_or_else(Err);
-
-			#[cfg(test)]
-			if let Some(control) = control.as_ref() {
-				control.worker_finished();
-			}
-
-			(content, admission)
-		});
-
-	#[cfg(test)]
-	super::tests::animation_submitted();
-
-	// a started worker runs to completion, but one still queued is dropped, so
-	// a cancelled request must not leave it to reach the pool
-	let abort = encode.abort_handle();
-
-	defer! {{ abort.abort(); }}
-
-	let (content, admission) = encode.await?;
-	let media = match content {
-		| Err(error) => Err(error),
-		| Ok(content) =>
-			self.store_encoded(mxc, dim, content, GIF, ANIMATED_NAME)
-				.await,
-	};
-
-	Ok(Output { media, admission })
+/// A source that animates leaves both variants stored at a size, so which
+/// one a lookup answers with must be stated rather than left to the order
+/// their keys fall in, which a remote row's own disposition decides. The
+/// other variant still answers when it is the only one there, so this
+/// orders the rows rather than refusing any of them.
+#[implement(Animate)]
+#[inline]
+#[must_use]
+pub fn prefers_type(self, content_type: Option<&str>) -> bool {
+	self.allowed() == content_type.is_some_and(declares_animation)
 }
 
-/// Encode a picture's frames as a GIF scaled to these dimensions.
+/// Returns true when this picture may answer the request.
 ///
-/// Both caps are settled before a frame is pulled, so neither the decode nor
-/// the encode can run past them: the sequence stops at the frame count or at
-/// what the pixel budget affords, and the animation loops short rather than the
-/// request failing. Fewer than two frames is no animation and answers an error.
-pub(in super::super) fn encode_frames(
-	bytes: &[u8],
-	dim: &Dim,
-	max_frames: usize,
-	budget: u64,
-) -> Result<Vec<u8>> {
-	let (frames, canvas) = source_frames(bytes, budget)?;
+/// The bytes decide, so a file cannot pass a request that forbade
+/// animation by declaring a content type that does not animate, and an
+/// APNG is caught despite being an `image/png` like every thumbnail.
+#[implement(Animate)]
+#[inline]
+#[must_use]
+pub fn accepts_picture(self, bytes: &[u8]) -> bool { self.allowed() || !animates(bytes) }
 
-	// every frame composites onto the whole canvas whatever region it updates,
-	// so what the budget affords is a count rather than a running total
-	let afforded = budget
-		.checked_div(canvas)
-		.and_then(|afforded| usize::try_from(afforded).ok())
-		.unwrap_or(max_frames);
+/// Returns true when a fetched picture may answer, walking it if nobody
+/// has.
+///
+/// Filing a row settles this on the way, and a redirect files none, so the
+/// walk that was skipped there happens here for the one caller that asks.
+#[implement(Animate)]
+#[must_use]
+pub fn accepts_fetched(self, fetched: &Fetched) -> bool {
+	self.accepts_walk(fetched.animates, &fetched.media.content)
+}
 
-	let mut content = Vec::new();
-	let mut encoder = GifEncoder::new_with_speed(&mut content, GIF_SPEED);
-	let mut count = 0_usize;
+/// Returns true when this picture may answer in a thumbnail's place.
+///
+/// Nothing can be derived from a picture that will not decode, so the
+/// choice is between the original and refusing media the server holds. Only
+/// a walk that settled on animation withholds it here, where
+/// [`Self::accepts_picture`] withholds anything it could not settle, since
+/// refusing every unreadable still would cost more than it buys.
+#[implement(Animate)]
+#[inline]
+#[must_use]
+pub fn accepts_fallback(self, bytes: &[u8]) -> bool {
+	self.allowed() || animated_type(bytes).is_none()
+}
 
-	encoder
-		.set_repeat(Repeat::Infinite)
-		.map_err(|error| err!(debug_warn!(?error, "Failed to set the GIF loop count.")))?;
+/// Returns true when this picture may answer, walking it if nobody has.
+///
+/// A caller already holding a walk of these bytes states what it settled
+/// rather than paying for a second one over them, and where none was taken
+/// this is [`Self::accepts_picture`] exactly.
+#[implement(Animate)]
+#[inline]
+#[must_use]
+pub(in super::super) fn accepts_walk(self, animates: Option<bool>, bytes: &[u8]) -> bool {
+	animates
+		.map_or_else(|| self.accepts_picture(bytes), |animates| self.accepts_animation(animates))
+}
 
-	for frame in frames.take(min(max_frames, afforded)) {
-		// a frame the source will not yield ends the sequence where it stands,
-		// as a cap does, rather than discarding the frames already encoded
-		let Ok(frame) = frame else {
-			break;
-		};
+/// Returns true when this picture may answer in a thumbnail's place,
+/// walking it if nobody has.
+///
+/// The settled-only rule of [`Self::accepts_fallback`] holds here too, so
+/// what the caller states is whether its walk *named* an animation rather
+/// than whether it withheld one.
+#[implement(Animate)]
+#[cfg(feature = "media_thumbnail")]
+#[inline]
+#[must_use]
+pub(in super::super) fn accepts_fallback_walk(self, names: Option<bool>, bytes: &[u8]) -> bool {
+	names.map_or_else(|| self.accepts_fallback(bytes), |names| self.allowed() || !names)
+}
 
-		let delay = frame.delay();
-		let scaled = thumbnail_generate(&DynamicImage::ImageRgba8(frame.into_buffer()), dim)?;
+/// Returns true when a picture may answer, given what a walk settled about
+/// it.
+///
+/// The walk that settles this is the same one picking the type a fetched
+/// row is filed under, so a caller already holding its answer states it
+/// here rather than reading the same bytes again through
+/// [`Self::accepts_picture`].
+#[implement(Animate)]
+#[inline]
+#[must_use]
+fn accepts_animation(self, animates: bool) -> bool { self.allowed() || !animates }
 
-		encoder
-			.encode_frame(Frame::from_parts(scaled.into_rgba8(), 0, 0, delay))
-			.map_err(|error| err!(debug_warn!(?error, "Failed to encode a frame.")))?;
-
-		count = count.saturating_add(1);
+impl From<Option<bool>> for Animate {
+	fn from(animated: Option<bool>) -> Self {
+		match animated {
+			| Some(true) => Self::Allowed,
+			| Some(false) | None => Self::Never,
+		}
 	}
-
-	// the encoder writes the trailer as it goes out of scope, so the picture is
-	// not complete until it has
-	drop(encoder);
-
-	if count < 2 {
-		return Err!(debug_warn!(%count, "Picture carries no frame sequence."));
-	}
-
-	Ok(content)
 }
 
-/// The frame sequence a picture carries, and what one frame of it costs.
-///
-/// Only the three containers MSC2705 names can hold a sequence, and two of them
-/// say in their own header whether they do, so a still answers an error here
-/// rather than yielding a lone frame for the caller to reject. Each decoder is
-/// built without limits of its own and composites a whole source canvas on its
-/// first advance, so the canvas is budgeted and the decoder told what it may
-/// allocate before it is ever advanced.
-fn source_frames(bytes: &[u8], budget: u64) -> Result<(Frames<'_>, u64)> {
-	let format = reader(bytes)?
-		.format()
-		.ok_or_else(|| err!(debug_warn!("Picture names no format.")))?;
-
-	let source = Cursor::new(bytes);
-	let failed = |error| err!(debug_warn!(?error, ?format, "Failed to read a frame sequence."));
-
-	let mut limits = Limits::no_limits();
-
-	limits.max_alloc = Some(budget.saturating_mul(BYTES_PER_PIXEL));
-
-	let frames = match format {
-		| ImageFormat::Gif => {
-			let mut decoder = GifDecoder::new(source).map_err(failed)?;
-			let canvas = canvas_pixels(&decoder, budget)?;
-
-			decoder.set_limits(limits).map_err(failed)?;
-
-			(decoder.into_frames(), canvas)
-		},
-		| ImageFormat::Png => {
-			let mut decoder = PngDecoder::new(source).map_err(failed)?;
-			let canvas = canvas_pixels(&decoder, budget)?;
-
-			if !decoder.is_apng().map_err(failed)? {
-				return Err!(debug_warn!("PNG carries a single frame."));
-			}
-
-			decoder.set_limits(limits).map_err(failed)?;
-
-			(decoder.apng().map_err(failed)?.into_frames(), canvas)
-		},
-		| ImageFormat::WebP => {
-			let mut decoder = WebPDecoder::new(source).map_err(failed)?;
-			let canvas = canvas_pixels(&decoder, budget)?;
-
-			if !decoder.has_animation() {
-				return Err!(debug_warn!("WebP carries a single frame."));
-			}
-
-			decoder.set_limits(limits).map_err(failed)?;
-
-			(decoder.into_frames(), canvas)
-		},
-		| _ => return Err!(debug_warn!(?format, "Format carries no frame sequence.")),
-	};
-
-	Ok(frames)
+impl From<Animate> for Option<bool> {
+	fn from(animate: Animate) -> Self { Some(animate.allowed()) }
 }
 
-/// Pixels one composited frame of this source occupies.
+/// Whether a declared content type names an animating container.
 ///
-/// A canvas the budget cannot afford even once is refused here, which is before
-/// the decoder has been advanced onto it and so before it has allocated one.
-fn canvas_pixels(decoder: &impl ImageDecoder, budget: u64) -> Result<u64> {
-	let (width, height) = decoder.dimensions();
-	let pixels = u64::from(width).saturating_mul(u64::from(height));
+/// This reads what an upload claimed rather than the picture itself, so it
+/// decides only between stored rows, where no picture is in hand. The reader
+/// that decides from the bytes is `sniff::animates`.
+fn declares_animation(content_type: &str) -> bool {
+	let essence = content_type
+		.split_once(';')
+		.map_or(content_type, |(essence, _)| essence)
+		.trim();
 
-	if pixels == 0 || pixels > budget {
-		return Err!(
-			debug_warn!(%width, %height, %budget, "Canvas is outside the pixel budget.")
-		);
-	}
-
-	Ok(pixels)
+	ANIMATED_TYPES
+		.iter()
+		.any(|kind| essence.eq_ignore_ascii_case(kind))
 }
